change int to double

+++ NEL_Aggr_API/Controllers/AuctionService.cs
@@ -310,12 +310,12 @@ namespace NEL_Agency_API.Controllers
                 if (maxBuyer != who)
                 {
                     maxBuyer = who;
-                    maxPrice = Convert.ToString(queryRes.Where(pItem => pItem["who"].ToString() == who && int.Parse(pItem["blockindex"].ToString()) <= int.Parse(item["blockindex"].ToString())).Sum(ppItem => int.Parse(ppItem["value"].ToString())));
+                    maxPrice = Convert.ToString(queryRes.Where(pItem => pItem["who"].ToString() == who && int.Parse(pItem["blockindex"].ToString()) <= int.Parse(item["blockindex"].ToString())).Sum(ppItem => double.Parse(ppItem["value"].ToString())));
                 }
                 long addPriceTime = blocktimeDict.GetValueOrDefault(Convert.ToString(item["blockindex"]));
                 return new JObject() { { "maxPrice", maxPrice }, { "maxBuyer", maxBuyer }, { "addPriceTime", addPriceTime } };
 
-            }).Where(p => Convert.ToString(p["maxPrice"]) != "0").OrderByDescending(p => p["addPriceTime"]).ThenByDescending(p => int.Parse(p["maxPrice"].ToString())).Skip(pageSize*(pageNum-1)).Take(pageSize).ToArray();
+            }).Where(p => Convert.ToString(p["maxPrice"]) != "0").OrderByDescending(p => p["addPriceTime"]).ThenByDescending(p => double.Parse(p["maxPrice"].ToString())).Skip(pageSize*(pageNum-1)).Take(pageSize).ToArray();
             long count = queryRes.Count;
             JObject res = new JObject();
             res.Add("list", new JArray() { arr });
